Refactor MCP input analysis helpers
