+++ crates/rulemorph_mcp/tests/stdio.rs
@@ -1315,6 +1315,38 @@ fn analyze_input_json_success() {
 }
 
 #[test]
+fn analyze_input_max_paths_limits_reported_paths() {
+    let mut server = McpServer::start();
+    initialize(&mut server);
+
+    let request = json!({
+        "jsonrpc": "2.0",
+        "id": 121,
+        "method": "tools/call",
+        "params": {
+            "name": "analyze_input",
+            "arguments": {
+                "input_json": {
+                    "id": 1,
+                    "name": "Ada",
+                    "active": true
+                },
+                "max_paths": 2
+            }
+        }
+    });
+
+    let response = server.send(&request);
+    let paths = response["result"]["meta"]["paths"]
+        .as_array()
+        .expect("paths array");
+    assert_eq!(paths.len(), 2);
+    assert_eq!(response["result"]["meta"]["summary"]["paths"], json!(2));
+
+    server.shutdown();
+}
+
+#[test]
 fn analyze_input_records_path_supports_quoted_segments() {
     let mut server = McpServer::start();
     initialize(&mut server);
